+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Search grids can define `fixed_parameters` which will always be applied to the query
+- Search grids can define a `selected_parameter_list` to only show those parameters to the user - even if the report definition has more
 ### Changed
 ### Fixed